Phase 4: Extract UI layer into src/ui/ package (~860 lines removed from app.py)
- New src/ui/__init__.py — package marker
- New src/ui/practice_tab.py — practice_word_from_audio wrapper, save_current_session,
  render_practice_interface, render_practice_results (with IPA diff display)
- New src/ui/story_tab.py — render_story_reader, render_full_story,
  render_scene_by_scene, render_scene_practice_mode
- New src/ui/statistics_tab.py — render_statistics_tab
- New src/ui/history_tab.py — load_history, save_history, render_history_tab
- src/app.py: import all ui modules; replace ~860 lines of inline definitions
  with thin import-based delegation; inline tab 3 and tab 4 blocks replaced
  with render_statistics_tab() and render_history_tab() calls
- New tests/test_ui.py — 15 smoke tests covering all exported names + save_history no-op

All 68 tests pass. App starts cleanly on port 8502 with no import errors.
