+++ src/app.py
@@ -86,6 +86,17 @@ from auth import (
     on_logout,
 )
 
+# UI tab modules (Phase 4 refactor)
+from ui.practice_tab import (
+    practice_word_from_audio,
+    save_current_session,
+    render_practice_interface,
+    render_practice_results,
+)
+from ui.story_tab import render_story_reader
+from ui.statistics_tab import render_statistics_tab
+from ui.history_tab import load_history, save_history, render_history_tab
+
 # Import API usage logger for cost tracking
 try:
     from api_usage_logger import log_api_call
@@ -329,48 +340,7 @@ with st.sidebar:
 # ============================================================================
 
 
-def load_history():
-    """Load practice history from database (if authenticated) or return empty list"""
-    if st.session_state.get('authenticated', False) and 'user' in st.session_state:
-        try:
-            user_id = st.session_state['user']['user_id']
-            language_code = st.session_state.get('language', 'Portuguese')
-            # Get recent progress from database
-            progress = app_mysql.get_user_progress(user_id, language_code, limit=100)
-
-            # Group practices by date into sessions for compatibility with old history format
-            from collections import defaultdict
-            sessions_by_date = defaultdict(list)
-
-            for p in progress:
-                date = p['practice_date'].date() if hasattr(p['practice_date'], 'date') else str(p['practice_date'])[:10]
-                sessions_by_date[date].append({
-                    "target": p['target_phrase'],
-                    "recognized": p['recognized_phrase'],
-                    "similarity": p['similarity_score'],
-                    "exact_match": p['perfect_match'],
-                    "correct_phonemes": p.get('target_phonemes', ''),
-                    "user_phonemes": p.get('user_phonemes', '')
-                })
-
-            # Convert to list of session objects
-            return [
-                {
-                    "date": str(date),
-                    "practices": practices
-                }
-                for date, practices in sorted(sessions_by_date.items(), reverse=True)
-            ]
-        except Exception as e:
-            st.warning(f"Could not load history from database: {e}")
-    return []
-
-
-def save_history(history: List[Dict]):
-    """Legacy function - history now saved immediately to database in check_pronunciation"""
-    # No-op: All practice results are saved to database immediately in check_pronunciation()
-    # This function kept for backward compatibility but does nothing
-    pass
+# load_history, save_history — now in ui/history_tab.py (imported above)
 
 
 def initialize_session_state():
@@ -466,765 +436,19 @@ def initialize_session_state():
 # compare_phonemes_edit_distance, compare_phonemes are all imported at top.
 
 
-def practice_word_from_audio(text: str, audio_bytes: bytes, settings: Dict):
-    """
-    Thin wrapper around scoring.practice.practice_word_from_audio that
-    handles session-state persistence and database saves.
+# practice_word_from_audio — now in ui/practice_tab.py (imported above)
+# save_current_session      — now in ui/practice_tab.py (imported above)
 
-    The core scoring logic lives in scoring/practice.py.
-    """
-    def _persist_result(result):
-        """Save result to session state and database."""
-        session_data = {
-            k: v for k, v in result.items()
-            if k not in ["user_audio_bytes", "user_audio_trimmed_bytes"]
-        }
-        st.session_state.current_sessions[st.session_state.language]["practices"].append({
-            "time": datetime.now().isoformat(),
-            **session_data
-        })
-        st.session_state.session_saved = False
 
-        if st.session_state.get('authenticated', False):
-            try:
-                user_id = st.session_state['user']['user_id']
-                app_mysql.save_practice(
-                    user_id=user_id,
-                    language_code=st.session_state.language,
-                    target_phrase=result['target'],
-                    recognized_phrase=result['recognized'],
-                    similarity_score=result['similarity'],
-                    perfect_match=result['exact_match'],
-                    target_phonemes=result['correct_ipa'],
-                    user_phonemes=result['user_ipa']
-                )
-            except Exception as e:
-                st.warning(f"⚠️ Could not save to database: {e}")
+# render_practice_interface — now in ui/practice_tab.py (imported above)
+# render_practice_results   — now in ui/practice_tab.py (imported above)
 
-    return _practice_word_from_audio_core(
-        text,
-        audio_bytes,
-        settings,
-        language=st.session_state.language,
-        on_result=_persist_result,
-    )
 
-
-def save_current_session():
-    """Save current session (for authenticated users, already saved to database)"""
-    current_session = st.session_state.current_sessions[st.session_state.language]
-    if current_session["practices"]:
-        if st.session_state.get('authenticated', False):
-            # For authenticated users, practices are already in database
-            st.success("✓ All practices saved to your account!")
-        else:
-            # For non-authenticated users (shouldn't happen now), append to local history
-            st.session_state.history.append(current_session)
-
-        # Reset current session for this language
-        st.session_state.current_sessions[st.session_state.language] = {
-            "date": datetime.now().isoformat(),
-            "practices": []
-        }
-        st.session_state.session_saved = True
-        st.rerun()
-    else:
-        st.warning("No practices in current session to save")
-
-
-def render_practice_interface(text, key_prefix="practice"):
-    """
-    Reusable practice interface component for audio playback, recording, and checking.
-
-    Args:
-        text: The phrase to practice
-        key_prefix: Unique prefix for widget keys to avoid collisions (default: "practice")
-
-    Returns:
-        None (handles UI rendering and result storage in session state)
-    """
-    if not text:
-        st.info("👆 Enter a word or phrase above to begin")
-        return
-
-    # Initialize mode-specific state variables
-    audio_key_name = f"{key_prefix}_audio_input_key"
-    if audio_key_name not in st.session_state:
-        st.session_state[audio_key_name] = 0
-
-    # Show target audio directly - one click to play
-    st.write("🎯 **Target pronunciation:**")
-    with st.spinner("Generating audio..."):
-        audio_bytes, audio_format = generate_target_audio(
-            text,
-            st.session_state.settings
-        )
-        st.audio(audio_bytes, format=audio_format, autoplay=False)
-
-    st.write("🎙️ **Now record your pronunciation:**")
-
-    # Streamlit's built-in audio input with dynamic key (unique per mode)
-    audio_data = st.audio_input("Click to record", key=f"{key_prefix}_audio_input_{st.session_state[audio_key_name]}")
-
-    # Show recording tip after the recording widget (mobile-friendly)
-    language_name = st.session_state.language
-    st.info(f"💡 Wait for the recording icon to turn red before speaking. The app will automatically trim silence and enforce {language_name} language detection.")
-
-    if audio_data:
-        st.write("▶️ **Your recording:**")
-        st.audio(audio_data, format='audio/wav')
-
-        # Always show both buttons when recording exists - critical for UX
-        col1, col2 = st.columns([1, 1])
-        with col1:
-            if st.button("✅ Check Pronunciation", key=f"{key_prefix}_submit_btn", type="primary"):
-                with st.spinner("Processing..."):
-                    result = practice_word_from_audio(
-                        text,
-                        audio_data.getvalue(),
-                        st.session_state.settings
-                    )
-                    if result:
-                        # Store result with mode-specific key
-                        st.session_state[f"{key_prefix}_last_result"] = result
-                        st.success("✓ Result processed successfully")
-                    else:
-                        st.error("❌ Processing failed - check terminal for errors")
-
-        with col2:
-            # CRITICAL: Always show Remove Recording button when audio exists
-            if st.button("🗑️ Remove Recording", key=f"{key_prefix}_clear_btn"):
-                # Clear the recording, results, and force widget reset
-                st.session_state[f"{key_prefix}_last_result"] = None
-                st.session_state[audio_key_name] += 1  # Change key to reset widget
-                st.rerun()
-
-
-def render_practice_results(result, key_prefix="practice"):
-    """
-    Reusable practice results display component.
-
-    Args:
-        result: Dictionary with practice results from practice_word_from_audio()
-        key_prefix: Unique prefix for widget keys to avoid collisions (default: "practice")
-    """
-    if not result:
-        return
-
-    st.markdown("---")
-    st.header("Results")
-
-    # Play celebration sounds based on score (only once per result)
-    import streamlit.components.v1 as components
-
-    # Track if sound has been played for this result
-    result_id = f"{result.get('target', '')}_{result.get('recognized', '')}_{result.get('similarity', 0)}"
-    if 'last_sound_played' not in st.session_state:
-        st.session_state.last_sound_played = None
-
-    should_play_sound = st.session_state.last_sound_played != result_id
-
-    if result["exact_match"]:
-        st.success("🎉 PERFECT MATCH! Well done!")
-        # Play perfect match bell sound (C major triad)
-        if should_play_sound:
-            st.session_state.last_sound_played = result_id
-            components.html(
-                """
-                <script>
-                (function () {
-                    try {
-                        const AudioContext = window.AudioContext || window.webkitAudioContext;
-                        if (!AudioContext) return;
-                        const ctx = new AudioContext();
-                        const now = ctx.currentTime;
-                        const freqs = [523.25, 659.25, 783.99]; // C5-E5-G5
-                        const baseGain = 0.3;
-                        const step = 0.15;
-                        const dur = 0.6;
-
-                        const play = () => {
-                            freqs.forEach((freq, i) => {
-                                const oscillator = ctx.createOscillator();
-                                const gainNode = ctx.createGain();
-                                oscillator.connect(gainNode);
-                                gainNode.connect(ctx.destination);
-                                oscillator.frequency.value = freq;
-                                oscillator.type = 'sine';
-                                gainNode.gain.setValueAtTime(baseGain, now + i * step);
-                                gainNode.gain.exponentialRampToValueAtTime(0.01, now + i * step + dur);
-                                oscillator.start(now + i * step);
-                                oscillator.stop(now + i * step + dur);
-                            });
-                        };
-
-                        if (ctx.resume) {
-                            Promise.resolve(ctx.resume()).then(play).catch(() => {});
-                        } else {
-                            play();
-                        }
-                    } catch (e) {
-                        // ignore
-                    }
-                })();
-                </script>
-                """,
-                height=0,
-            )
-    elif result['similarity'] >= 0.90:
-        # High score but not perfect: gentle encouraging sound
-        st.success(f"✨ Excellent! {result['similarity']:.1%} - Almost perfect!")
-        if should_play_sound:
-            st.session_state.last_sound_played = result_id
-            components.html(
-                """
-                <script>
-                (function () {
-                    try {
-                        const AudioContext = window.AudioContext || window.webkitAudioContext;
-                        if (!AudioContext) return;
-                        const ctx = new AudioContext();
-                        const now = ctx.currentTime;
-                        const freqs = [440, 493.88, 523.25]; // A4-B4-C5
-                        const baseGain = 0.15;
-                        const step = 0.12;
-                        const dur = 0.4;
-
-                        const play = () => {
-                            freqs.forEach((freq, i) => {
-                                const oscillator = ctx.createOscillator();
-                                const gainNode = ctx.createGain();
-                                oscillator.connect(gainNode);
-                                gainNode.connect(ctx.destination);
-                                oscillator.frequency.value = freq;
-                                oscillator.type = 'sine';
-                                gainNode.gain.setValueAtTime(baseGain, now + i * step);
-                                gainNode.gain.exponentialRampToValueAtTime(0.01, now + i * step + dur);
-                                oscillator.start(now + i * step);
-                                oscillator.stop(now + i * step + dur);
-                            });
-                        };
-
-                        if (ctx.resume) {
-                            Promise.resolve(ctx.resume()).then(play).catch(() => {});
-                        } else {
-                            play();
-                        }
-                    } catch (e) {
-                        // ignore
-                    }
-                })();
-                </script>
-                """,
-                height=0,
-            )
-    else:
-        score_col1, score_col2 = st.columns([2, 1])
-        with score_col1:
-            st.info(f"📊 Score: {result['similarity']:.1%}")
-        with score_col2:
-            if result.get('edit_distance') is not None:
-                st.metric("Edit Distance", result['edit_distance'],
-                        help="Number of edits needed to match target")
-
-    col1, col2 = st.columns(2)
-    with col1:
-        st.subheader("Target")
-        st.write(f"**Text:** {result['target']}")
-        if result.get('correct_ipa'):
-            st.markdown(f"**IPA:** {format_ipa(result['correct_ipa'])}", unsafe_allow_html=True)
-
-        # Show target audio directly
-        tts_label = "Google TTS" if st.session_state.settings.get('tts_engine', 'gtts') == 'gtts' else "eSpeak"
-        st.write(f"🔊 **{tts_label}:**")
-        audio_bytes, audio_format = generate_target_audio(result['target'], st.session_state.settings)
-        st.audio(audio_bytes, format=audio_format)
-
-    with col2:
-        st.subheader("Your Pronunciation")
-        st.write(f"**Recognized:** {result['recognized']}")
-        if result.get('user_ipa'):
-            st.markdown(f"**IPA:** {format_ipa(result['user_ipa'])}", unsafe_allow_html=True)
-
-        # Show comparison note
-        # Normalize text by removing punctuation for comparison
-        import string
-        target_clean = result['target'].lower().translate(str.maketrans('', '', string.punctuation))
-        recognized_clean = result['recognized'].translate(str.maketrans('', '', string.punctuation))
-
-        # Get normalized phonemes for comparison (the source of truth for pronunciation)
-        correct_phonemes_no_space = result.get('correct_phonemes_normalized') or normalize_for_phoneme_scoring(result.get('correct_phonemes', ''))
-        user_phonemes_no_space = result.get('user_phonemes_normalized') or normalize_for_phoneme_scoring(result.get('user_phonemes', ''))
-
-        # Check if phonemes match (pronunciation is correct)
-        phonemes_match = correct_phonemes_no_space == user_phonemes_no_space
-        text_matches = target_clean == recognized_clean
-        score_is_high = result['similarity'] >= 0.95
-
-        # Display appropriate message based on phoneme match (source of truth)
-        if phonemes_match:
-            # Phonemes match perfectly - pronunciation is correct
-            if not text_matches:
-                st.success("✅ Perfect pronunciation! (Text punctuation/formatting differs)")
-            else:
-                st.success("✅ Phonemes match perfectly")
-        elif score_is_high and not text_matches:
-            # High score but text differs slightly
-            st.info("ℹ️ Excellent pronunciation! (Minor text differences)")
-        elif not text_matches and not score_is_high:
-            # Both text and pronunciation differ - likely wrong word or unclear speech
-            st.warning("⚠️ Different words recognized - try speaking more clearly")
-
-    # Close the two-column layout before the detailed analysis
-    # Show detailed phoneme analysis (works with edit distance!) - full width
-    if st.checkbox("🔍 Show detailed phoneme analysis", key=f"{key_prefix}_show_detail"):
-        st.markdown("---")
-
-        st.markdown("#### Phoneme Analysis")
-        st.write(f"**Algorithm:** {st.session_state.settings.get('comparison_algorithm', 'edit_distance')}")
-
-        if result.get('edit_distance') is not None:
-            st.write(f"**Edit Distance:** {result['edit_distance']} edit(s) needed")
-
-        # Primary display: IPA (user-friendly)
-        correct_ipa = result.get('correct_ipa', '') or ''
-        user_ipa = result.get('user_ipa', '') or ''
-
-        st.write("**IPA (from eSpeak, for readability):**")
-        col_a, col_b = st.columns(2)
-        with col_a:
-            if correct_ipa:
-                st.markdown(format_ipa(correct_ipa), unsafe_allow_html=True)
-            else:
-                st.write("(no IPA available)")
-            st.caption("Target")
-        with col_b:
-            if user_ipa:
-                st.markdown(format_ipa(user_ipa), unsafe_allow_html=True)
-            else:
-                st.write("(no IPA available)")
-            st.caption("Your Pronunciation")
-
-        # Comparison for display (whitespace ignored)
-        target_ipa_no_space = "".join(correct_ipa.split())
-        user_ipa_no_space = "".join(user_ipa.split())
-
-        st.write("**Detailed IPA comparison (whitespace ignored):**")
-        if target_ipa_no_space and target_ipa_no_space == user_ipa_no_space:
-            st.success("🎯 IPA is identical!")
-        elif target_ipa_no_space or user_ipa_no_space:
-            from difflib import SequenceMatcher
-            import html as _html
-
-            def _colorize_diff(target: str, user: str) -> tuple[str, str]:
-                # replace: light blue, insert: light green, delete: light pink.
-                matcher_local = SequenceMatcher(None, target, user)
-                target_chunks: list[str] = []
-                user_chunks: list[str] = []
-
-                for tag, i1, i2, j1, j2 in matcher_local.get_opcodes():
-                    t_seg = target[i1:i2]
-                    u_seg = user[j1:j2]
-
-                    if tag == 'equal':
-                        target_chunks.append(_html.escape(t_seg))
-                        user_chunks.append(_html.escape(u_seg))
-                    elif tag == 'replace':
-                        target_chunks.append(f'<span style="background-color: #ADD8E6; padding: 0 2px;">{_html.escape(t_seg)}</span>')
-                        user_chunks.append(f'<span style="background-color: #ADD8E6; padding: 0 2px;">{_html.escape(u_seg)}</span>')
-                    elif tag == 'insert':
-                        target_chunks.append(f'<span style="background-color: #90EE90; padding: 0 2px;">{_html.escape("·" * len(u_seg))}</span>')
-                        user_chunks.append(f'<span style="background-color: #90EE90; padding: 0 2px;">{_html.escape(u_seg)}</span>')
-                    elif tag == 'delete':
-                        target_chunks.append(f'<span style="background-color: #FFB6C6; padding: 0 2px;">{_html.escape(t_seg)}</span>')
-                        user_chunks.append(f'<span style="background-color: #FFB6C6; padding: 0 2px;">{_html.escape("·" * len(t_seg))}</span>')
-
-                return ''.join(target_chunks), ''.join(user_chunks)
-
-            matcher = SequenceMatcher(None, target_ipa_no_space, user_ipa_no_space)
-            operations = matcher.get_opcodes()
-            substitutions = [op for op in operations if op[0] == 'replace']
-            insertions = [op for op in operations if op[0] == 'insert']
-            deletions = [op for op in operations if op[0] == 'delete']
-            matches = sum(i2 - i1 for tag, i1, i2, j1, j2 in operations if tag == 'equal')
-            st.write(f"**Operations:** {matches} matches, {len(substitutions)} substitutions, {len(insertions)} insertions, {len(deletions)} deletions")
-
-            target_html, user_html = _colorize_diff(target_ipa_no_space, user_ipa_no_space)
-            mono_wrap_start = '<div style="font-family: ui-monospace, SFMono-Regular, Menlo, Monaco, Consolas, \'Liberation Mono\', \'Courier New\', monospace; white-space: pre-wrap;">'
-            mono_wrap_end = '</div>'
-
-            col_t, col_u = st.columns(2)
-            with col_t:
-                st.markdown(mono_wrap_start + target_html + mono_wrap_end, unsafe_allow_html=True)
-                st.caption("Target (normalized) - substitutions/insertions/deletions highlighted")
-            with col_u:
-                st.markdown(mono_wrap_start + user_html + mono_wrap_end, unsafe_allow_html=True)
-                st.caption("Your Pronunciation (normalized) - substitutions/insertions/deletions highlighted")
-        else:
-            st.info("No IPA available for detailed comparison.")
-
-        # Technical: show the eSpeak phoneme codes used for scoring
-        with st.expander("Technical: eSpeak phoneme codes used for scoring", expanded=False):
-            st.write("**eIPA (eSpeak -x) with word spacing:**")
-            col_xa, col_xb = st.columns(2)
-            with col_xa:
-                st.code(result.get('correct_phonemes', ''), language=None)
-                st.caption("Target")
-            with col_xb:
-                st.code(result.get('user_phonemes', ''), language=None)
-                st.caption("Your Pronunciation")
-
-            target_phonemes_no_space = result.get('correct_phonemes_normalized', '')
-            user_phonemes_no_space = result.get('user_phonemes_normalized', '')
-            st.write("**eIPA used for scoring (whitespace removed):**")
-            col_n1, col_n2 = st.columns(2)
-            with col_n1:
-                st.code(target_phonemes_no_space, language=None)
-                st.caption("Target (normalized)")
-            with col_n2:
-                st.code(user_phonemes_no_space, language=None)
-                st.caption("Your Pronunciation (normalized)")
-
-
-def render_scene_practice_mode(scenes_dir):
-    """
-    Story practice mode - practice pronunciation of story phrases scene by scene.
-
-    Args:
-        scenes_dir: Path to the directory containing scene JSON files
-    """
-    if not scenes_dir.exists():
-        st.warning("Story scenes not found. Please ensure story-scenes-json/ exists.")
-        return
-
-    # Get all scene files
-    scene_files = sorted(scenes_dir.glob("scene-*.json"))
-
-    if not scene_files:
-        st.warning("No scene files found in the story-scenes-json directory.")
-        return
-
-    # Initialize session state for story practice
-    if 'story_practice_scene_file' not in st.session_state:
-        st.session_state.story_practice_scene_file = str(scene_files[0])
-    if 'story_practice_index' not in st.session_state:
-        st.session_state.story_practice_index = 0
-
-    # Create scene selector with friendly names
-    scene_options = {}
-    for scene_file in scene_files:
-        # Extract scene number and title from filename
-        parts = scene_file.stem.split('-', 2)
-        if len(parts) >= 3:
-            scene_num = parts[1]
-            scene_title = parts[2].replace('-', ' ').title()
-            display_name = f"Scene {scene_num}: {scene_title}"
-        else:
-            display_name = scene_file.stem
-
-        scene_options[display_name] = str(scene_file)
-
-    # Scene selector
-    selected_scene_display = st.selectbox(
-        "Select a scene to practice:",
-        list(scene_options.keys()),
-        index=list(scene_options.values()).index(st.session_state.story_practice_scene_file)
-            if st.session_state.story_practice_scene_file in scene_options.values() else 0,
-        help="Choose a scene from Sophie & Lucas's adventure",
-        key="story_practice_scene_select"
-    )
-
-    selected_scene_path = scene_options[selected_scene_display]
-
-    # If scene changed, reset index
-    if selected_scene_path != st.session_state.story_practice_scene_file:
-        st.session_state.story_practice_scene_file = selected_scene_path
-        st.session_state.story_practice_index = 0
-        st.session_state.story_last_result = None
-        st.rerun()
-
-    # Load the scene
-    try:
-        with open(selected_scene_path, 'r', encoding='utf-8') as f:
-            scene_data = json.load(f)
-
-        # All story scenes now use Format 2: {"lang": [...], "scene_number": 1, "scene_title": "..."}
-        if not isinstance(scene_data, dict):
-            st.error(f"Invalid scene format - expected dict, got {type(scene_data).__name__}")
-            return
-
-        # Get language key (pt, fr, de, nl, it, es)
-        lang_keys = [k for k in scene_data.keys() if k not in ['scene_number', 'scene_title']]
-        if not lang_keys:
-            st.error("Invalid scene data - no language key found")
-            return
-
-        lang_key = lang_keys[0]
-        phrases = scene_data[lang_key]
-
-        if not phrases:
-            st.warning("No phrases found in this scene.")
-            return
-
-        # Navigation and progress
-        total_phrases = len(phrases)
-        current_idx = st.session_state.story_practice_index
-
-        # Keep index in bounds
-        if current_idx >= total_phrases:
-            st.session_state.story_practice_index = 0
-            current_idx = 0
-
-        current_phrase_obj = phrases[current_idx]
-        current_phrase = current_phrase_obj.get(lang_key, '')
-        phrase_translation = current_phrase_obj.get('english')
-        phrase_ipa = current_phrase_obj.get('ipa')
-
-        st.subheader(selected_scene_display)
-
-        # Progress bar
-        progress = (current_idx + 1) / total_phrases
-        st.progress(progress, text=f"Phrase {current_idx + 1} of {total_phrases}")
-
-        # Navigation buttons
-        col1, col2, col3 = st.columns([1, 1, 2])
-        with col1:
-            if st.button("⬅️ Previous", disabled=(current_idx == 0), key="story_prev"):
-                st.session_state.story_practice_index -= 1
-                st.session_state.story_last_result = None
-                st.rerun()
-        with col2:
-            if st.button("Next ➡️", disabled=(current_idx >= total_phrases - 1), key="story_next"):
-                st.session_state.story_practice_index += 1
-                st.session_state.story_last_result = None
-                st.rerun()
-        with col3:
-            st.caption(f"💡 Navigate through {total_phrases} phrases in this scene")
-
-        st.markdown("---")
-
-        # Display current phrase
-        if phrase_translation or phrase_ipa:
-            with st.expander("📖 Translation & Reference", expanded=False):
-                if phrase_translation:
-                    st.markdown(f"**🇬🇧 English:** {phrase_translation}")
-                if phrase_ipa:
-                    st.markdown(f"**📚 Reference IPA:** {format_ipa(phrase_ipa)}", unsafe_allow_html=True)
-                    st.caption("Compare with eSpeak IPA generated below")
-
-        st.markdown(f"#### 🎯 **{current_phrase}**")
-
-        # Practice interface with unique key prefix for story mode
-        render_practice_interface(current_phrase, key_prefix="story")
-
-        # Show results
-        if st.session_state.get('story_last_result'):
-            render_practice_results(st.session_state.story_last_result, key_prefix="story")
-
-    except json.JSONDecodeError as e:
-        st.error(f"Error parsing scene file: {e}")
-    except Exception as e:
-        st.error(f"Error loading scene: {e}")
-
-
-def render_story_reader():
-    """
-    Story Reader tab - Read stories in various formats
-    Modular design allows easy UX refactoring
-    Language-aware: displays story for current target language
-    """
-    try:
-        # Get material language from session state (affects story display)
-        lang_code = st.session_state.get('material_language', 'fr')
-
-        # Story titles and paths per language
-        story_config = {
-            'pt': {'title': 'Sophie & Lucas: Uma Jornada aos Alpes', 'setting': 'Brazil'},
-            'fr': {'title': 'Sophie & Lucas: A Journey to the Alps', 'setting': 'French Alps'},
-            'nl': {'title': 'Sophie & Lucas: Een Reis naar de Alpen', 'setting': 'Netherlands'},
-            'de': {'title': 'Sophie & Lucas: Eine Reise in die Alpen', 'setting': 'Black Forest/Alps'},
-            'it': {'title': 'Sophie & Lucas: Un Viaggio sulle Alpi', 'setting': 'Italian Dolomites'},
-            'es': {'title': 'Sophie & Lucas: Un Viaje a Sierra Nevada', 'setting': 'Sierra Nevada, Spain'}
-        }
-
-        # Check if story materials exist for this language
-        story_md_path = Path(f"language_materials/{lang_code}/story.md")
-        story_scenes_dir = Path(f"language_materials/{lang_code}/story-scenes-json")
-
-        config = story_config.get(lang_code, {'title': 'Story', 'setting': 'Unknown'})
-        st.header(f"📖 {config['title']}")
-
-        # Check what story materials are available
-        has_full_story = story_md_path.exists()
-        has_scenes = story_scenes_dir.exists() and list(story_scenes_dir.glob("scene-*.json"))
-
-        if not has_full_story and not has_scenes:
-            st.warning("Story materials not found. Please ensure story files exist for this language.")
-            return
-
-        # Story mode selector - show only available modes
-        available_modes = []
-        if has_full_story:
-            available_modes.append("📄 Full Story")
-        if has_scenes:
-            available_modes.extend(["🎬 Scene by Scene", "🎙️ Practice Mode"])
-
-        # Preserve story_mode across tab switches
-        # Check if we have a saved preference
-        saved_mode = st.session_state.get('_story_mode_preference')
-
-        # Calculate index: use saved mode if it's still available, otherwise default to Scene by Scene
-        default_idx = 0
-        if saved_mode and saved_mode in available_modes:
-            default_idx = available_modes.index(saved_mode)
-        elif "🎬 Scene by Scene" in available_modes:
-            default_idx = available_modes.index("🎬 Scene by Scene")
-
-        story_mode = st.radio(
-            "Choose reading mode:",
-            available_modes,
-            index=default_idx,
-            horizontal=True,
-            key='story_mode',
-            help="Read the complete story, explore individual scenes, or practice pronunciation"
-        )
-
-        # Save preference for next time
-        st.session_state._story_mode_preference = story_mode
-
-        if story_mode == "📄 Full Story":
-            render_full_story(story_md_path)
-        elif story_mode == "🎬 Scene by Scene":
-            render_scene_by_scene(story_scenes_dir, lang_code)
-        elif story_mode == "🎙️ Practice Mode":
-            render_scene_practice_mode(story_scenes_dir)
-
-    except Exception as e:
-        st.error(f"❌ Story Reader Error: {e}")
-        import traceback
-        st.error(f"```\n{traceback.format_exc()}\n```")
-
-
-def render_full_story(story_path):
-    """Render the complete story from markdown file"""
-    try:
-        with open(story_path, 'r', encoding='utf-8') as f:
-            story_content = f.read()
-
-        # Display the story
-        st.markdown(story_content, unsafe_allow_html=False)
-
-    except Exception as e:
-        st.error(f"Error loading story: {e}")
-
-
-def render_scene_by_scene(scenes_dir, lang_code):
-    """Render individual scenes with target language text and English translations"""
-    if not scenes_dir.exists():
-        st.warning(f"Story scenes not found. Please ensure `language_materials/{lang_code}/story-scenes-json/` exists.")
-        return
-
-    # Get all scene files
-    scene_files = sorted(scenes_dir.glob("scene-*.json"))
-
-    if not scene_files:
-        st.warning("No scene files found in the story-scenes-json directory.")
-        return
-
-    # Create scene selector with friendly names
-    scene_options = {}
-    for scene_file in scene_files:
-        # Extract scene number and title from filename
-        # e.g., "scene-01-le-café-du-matin.json" -> "Scene 1: Le Café du Matin"
-        parts = scene_file.stem.split('-', 2)
-        if len(parts) >= 3:
-            scene_num = parts[1]
-            scene_title = parts[2].replace('-', ' ').title()
-            display_name = f"Scene {scene_num}: {scene_title}"
-        else:
-            display_name = scene_file.stem
-
-        scene_options[display_name] = scene_file
-
-    # Scene selector
-    selected_scene = st.selectbox(
-        "Select a scene to read:",
-        list(scene_options.keys()),
-        help="Choose a scene from Sophie & Lucas's adventure"
-    )
-
-    scene_file = scene_options[selected_scene]
-
-    # Load and display the scene
-    try:
-        with open(scene_file, 'r', encoding='utf-8') as f:
-            scene_data = json.load(f)
-
-        # All story scenes now use Format 2: {"lang": [...], "scene_number": 1, "scene_title": "..."}
-        if not isinstance(scene_data, dict):
-            st.error(f"Invalid scene format - expected dict, got {type(scene_data).__name__}")
-            return
-
-        # Get language key (pt, fr, de, nl, it, es)
-        lang_keys = [k for k in scene_data.keys() if k not in ['scene_number', 'scene_title']]
-        if not lang_keys:
-            st.error("Invalid scene data - no language key found")
-            return
-
-        lang_key = lang_keys[0]
-        phrases = scene_data[lang_key]
-
-        st.subheader(selected_scene)
-        st.caption(f"📊 {len(phrases)} phrases in this scene")
-
-        # Display options
-        col1, col2 = st.columns([3, 1])
-        with col1:
-            source_lang = st.session_state.source_language
-            show_translations = st.checkbox(f"Show {source_lang} translations", value=False)
-        with col2:
-            show_ipa = st.checkbox("Show IPA", value=False)
-
-        st.divider()
-
-        # Display each phrase
-        for i, phrase in enumerate(phrases, 1):
-            # Get text in the target language (french, pt, etc.)
-            target_text = phrase.get(lang_key, '')
-            source_lang = st.session_state.source_language
-            target_lang = st.session_state.target_language
-
-            if source_lang == "English":
-                translation_text = phrase.get('english', '[Translation missing]')
-            else:
-                translation_text = get_translation_from_llm(target_text, target_lang, source_lang)
-
-            ipa_text = phrase.get('ipa', '')
-
-            # Target language text (always shown)
-            st.markdown(f"**{i}.** {target_text}")
-
-            # Optional: Source-language translation
-            if show_translations and translation_text and not translation_text.startswith('[error'):
-                st.markdown(f"   *{translation_text}*")
-
-            # Optional: IPA
-            if show_ipa and ipa_text:
-                st.markdown(f"   🔊 {format_ipa(ipa_text)}", unsafe_allow_html=True)
-
-            # Add spacing between phrases
-            if i < len(scene_data):
-                st.markdown("")  # Small gap
-
-        # Practice transition
-        st.divider()
-        st.info("✏️ **Ready to practice?** Go to the **🎯 Quick Practice** tab and load this scene from the Built-in Library → French → Story Scenes.")
-
-    except json.JSONDecodeError as e:
-        st.error(f"Error parsing scene file: {e}")
-    except Exception as e:
-        st.error(f"Error loading scene: {e}")
-
+# render_practice_results     — now in ui/practice_tab.py (imported above)
+# render_scene_practice_mode  — now in ui/story_tab.py (imported above)
+# render_story_reader          — now in ui/story_tab.py (imported above)
+# render_full_story            — now in ui/story_tab.py (imported above)
+# render_scene_by_scene        — now in ui/story_tab.py (imported above)
 
 def main():
     """Main Streamlit app"""
@@ -2391,71 +1615,11 @@ def main():
 
     # Tab 3: Statistics
     elif selected_tab_index == 2:
-        st.header("📊 Practice Statistics")
-
-        # Current session stats
-        current_session = st.session_state.current_sessions[st.session_state.language]
-        if current_session["practices"]:
-            st.subheader("🔵 Current Session")
-            practices = current_session["practices"]
-            perfect = sum(1 for p in practices if p.get("exact_match", False))
-            avg_sim = sum(p["similarity"] for p in practices) / len(practices)
-
-            col1, col2, col3 = st.columns(3)
-            col1.metric("Practices", len(practices))
-            col2.metric("Perfect", f"{perfect} ({perfect/len(practices):.1%})")
-            col3.metric("Avg Similarity", f"{avg_sim:.1%}")
-
-        # Overall stats - from database for authenticated users
-        if st.session_state.get('authenticated', False):
-            st.subheader("📈 All Time")
-            try:
-                user_id = st.session_state['user']['user_id']
-                stats = app_mysql.get_user_stats(user_id, st.session_state.language)
-
-                if stats and stats['total'] > 0:
-                    col1, col2, col3, col4 = st.columns(4)
-                    col1.metric("Total Practices", stats['total'])
-                    col2.metric("Total Perfect", f"{stats['perfect_count']} ({stats['perfect_count']/stats['total']:.1%})")
-                    col3.metric("Overall Avg", f"{stats['avg_score']:.1%}")
-                    col4.metric("Recent Avg (last 10)", f"{stats['recent_avg']:.1%}")
-                else:
-                    st.info("No practice history yet. Start practicing!")
-            except Exception as e:
-                st.error(f"Could not load stats: {e}")
-        else:
-            st.info("No practice history yet. Start practicing!")
+        render_statistics_tab()
 
     # Tab 4: History
     elif selected_tab_index == 3:
-        st.header("📜 Session History")
-
-        # Reload history from database when viewing this tab
-        st.session_state.history = load_history()
-
-        if not st.session_state.history:
-            st.info("No previous sessions")
-        else:
-            # History is already sorted newest first from load_history()
-            for i, session in enumerate(st.session_state.history[-10:], 1):
-                date = session["date"][:10]
-                count = len(session["practices"])
-                perfect = sum(1 for p in session["practices"] if p.get("exact_match", False))
-
-                with st.expander(f"{date} - {count} practices ({perfect} perfect)"):
-                    for j, practice in enumerate(session["practices"], 1):
-                        status = "✅" if practice.get("exact_match", False) else f"📊 {practice.get('similarity', 0):.1%}"
-
-                        st.markdown(f"**{j}. {status}**")
-                        col1, col2 = st.columns(2)
-
-                        with col1:
-                            st.write("Target:", practice.get('target', 'N/A'))
-
-                        with col2:
-                            st.write("Recognized:", practice.get('recognized', 'N/A'))
-
-                        st.markdown("---")
+        render_history_tab()
 
     # CCS Testing: Extract app state after UI renders (if testing enabled)
     if CCS_AVAILABLE and st.session_state.ccs_test.enabled:
