+++ src/ui/__init__.py
@@ -1,0 +1,1 @@
+# ui — tab-level rendering modules
